Moved get_benchmark_results to views_data.py
get_benchmark_results now relies on calling function to catch any
ObjectNotFound exception.
